Changes to error handling and signup.
